requirements.txt
diffusers
torch
torchvision
pipeline
transformers
accelerate
safetensors
spaces
peft
gradio
pillow